+++ dbt/include/firebolt/macros/adapters.sql
@@ -178,10 +178,10 @@
 {% endmacro %}
 
 
-{% macro firebolt__list_relations_without_caching(schema_relation) %}
+{% macro firebolt__list_relations_without_caching(relation) %}
   {# Return all views and tables as agate table.
      Args:
-       schema_relation (dict): Contains values for database and schema.
+       relation (dict): Contains values for database and schema.
 
      dbt has a relations cache. Using this macro will list all
      the relations in the current schema using a direct DB query,
@@ -190,15 +190,15 @@
   #}
   {% call statement('list_tables_without_caching', fetch_result=True) %}
 
-      SELECT '{{ schema_relation.database }}' AS "database",
+      SELECT '{{ relation.database }}' AS "database",
              table_name AS "name",
-             '{{ schema_relation.schema }}' AS "schema",
+             '{{ relation.schema }}' AS "schema",
              'table' AS type
         FROM information_schema.tables
       UNION
-      SELECT '{{ schema_relation.database }}' AS "database",
+      SELECT '{{ relation.database }}' AS "database",
              table_name AS "name",
-             '{{ schema_relation.schema }}' AS "schema",
+             '{{ relation.schema }}' AS "schema",
              'view' AS type
         FROM information_schema.views
   {% endcall %}
@@ -207,7 +207,7 @@
 {% endmacro %}
 
 
-{% macro firebolt__create_table_as(temporary, relation, sql) -%}
+{% macro firebolt__create_table_as(temporary, relation, ctas_sql) -%}
   {# Create table using CTAS
      Args:
       temporary (bool): Unused, included so macro signature matches
@@ -215,24 +215,34 @@
       relation (dbt relation/dict):
   #}
   {%- set table_type = config.get('table_type', default='dimension') | upper -%}
-  {%- set primary_index = config.get('primary_index') %}
+  {%- set primary_index = config.get('primary_index') -%}
+  {%- set incremental_strategy = config.get('incremental_strategy') -%}
+  {%- set partitions = config.get('partition_by') %}
 
   CREATE {{ table_type }} TABLE IF NOT EXISTS {{ relation }}
   {%- if primary_index %}
-      PRIMARY INDEX
-      {% if primary_index is iterable and primary_index is not string -%}
-          {{ primary_index | join(', ') }}
-      {%- else -%}
-          {{ primary_index }}
-      {%- endif -%}
-  {% endif %}
+  PRIMARY INDEX
+    {% if primary_index is iterable and primary_index is not string %}
+      {{ primary_index | join(', ') }}
+    {%- else -%}
+      {{ primary_index }}
+    {%- endif -%}
+  {%- endif -%}
+  {% if partitions %}
+  PARTITION BY
+    {% if partitions is iterable and partitions is not string %}
+      {{ partitions | join(', ') }}
+    {%- else -%}
+      {{ partitions }}
+    {%- endif -%}
+  {%- endif  %}
   AS (
-      {{ sql }}
+    {{ ctas_sql }}
   )
 {% endmacro %}
 
 
-{% macro firebolt__create_view_as(relation, sql) %}
+{% macro firebolt__create_view_as(relation, ctas_sql) %}
   {#-
   Return SQL string to create view.
      Args:
@@ -241,7 +251,7 @@
   #}
 
     CREATE OR REPLACE VIEW {{ relation.identifier }} AS (
-        {{ sql }}
+      {{ ctas_sql }}
     )
 {% endmacro %}